chore: update folder structure in the playback package
