+++ package.json
@@ -35,7 +35,7 @@
     "@nx/js": "20.3.1",
     "@size-limit/preset-big-lib": "^11.1.6",
     "@types/eslint__js": "^8.42.3",
-    "@types/react": "^19.0.6",
+    "@types/react": "^19.0.7",
     "@types/react-dom": "^19.0.3",
     "@typestrong/ts-mockito": "^2.7.12",
     "@vitejs/plugin-react": "^4.3.4",
@@ -45,7 +45,7 @@
     "eslint": "^9.18.0",
     "eslint-config-prettier": "^10.0.1",
     "eslint-plugin-jsdoc": "^50.6.1",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-prettier": "^5.2.2",
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.18",
     "eslint-plugin-unicorn": "^56.0.1",
