--- packages/playback/src/entry-points/main-only-core.ts
@@ -1,1 +1,0 @@
-export { Player } from '../lib/player/main-thread-only/main-thread-only-player';
