+++ packages/playback/src/entry-points/player.ts
@@ -1,0 +1,1 @@
+export { Player } from '../lib/player/player';
