--- packages/playback/src/lib/player/main-thread-with-worker/worker/worker-script.ts
@@ -1,4 +1,0 @@
-import { PipelineLoaderFactoryStorage } from '../../base/pipeline-loader-factory-storage';
-import { WorkerBridge } from './worker-bridge';
-
-WorkerBridge.create(new PipelineLoaderFactoryStorage());
