+++ packages/playback/src/lib/worker/worker-script.ts
@@ -1,0 +1,4 @@
+import { PipelineLoaderFactoryStorage } from '../utils/pipeline-loader-factory-storage';
+import { WorkerBridge } from './worker-bridge';
+
+WorkerBridge.create(new PipelineLoaderFactoryStorage());
